Update API-words-create: nullable furigana and insert default image

+++ app/Http/Controllers/WordController.php
@@ -70,7 +70,8 @@ class WordController extends Controller
                         "explanation_of_meaning" => $x["explanation_of_meaning"],
                         "example" => $x["example"],
                         "example_meaning" => $x["example_meaning"],
-                        "image" => $x["image"]
+                        "image" => "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTNET4coNATuFn3TwH9_dn5FvMp2hjKPANHGA&usqp=CAU"
+                        // "image" => $x["image"]
                     );
                     $meaning = Meaning::create($data_meaning);
                     foreach($x["tags_id"] as $tag_id){
@@ -86,6 +87,17 @@ class WordController extends Controller
                             'word1_id' => $word->id,
                             'word2_id' => $word2_id,
                             'relation_type' => 1
+                        ]);
+                }
+            }
+            if (isset($antonym)) {
+                foreach ($antonym as $value) {
+                    $word2_id = Word::where('word', $value)->first()->id;
+                    DB::table('word_relations')
+                        ->updateOrInsert([
+                            'word1_id' => $word->id,
+                            'word2_id' => $word2_id,
+                            'relation_type' => 0
                         ]);
                 }
             }
@@ -147,36 +159,64 @@ class WordController extends Controller
         if ($validator->fails()) {
             return response()->json($validator->errors(), 422);
         }
+        DB::beginTransaction();
         try {
             $word = Word::find($id);
             $word->update($data);
 
-            // Delete old data related
-            $meaning = Meaning::where('word_id',$id);
-            foreach($meaning as $m){
-                $m->tags()->detach();
-            }
-            Meaning::where('word_id',$id)->delete();
+            if (isset($arr_meanings)) {
+                // Delete old data related
+                $meaning = Meaning::where('word_id',$id);
+                foreach($meaning as $m){
+                    $m->tags()->detach();
+                }
+                Meaning::where('word_id',$id)->delete();
 
-            // Create new data related
-            foreach($arr_meanings as $x){
-            
-                $data_meaning = array(
-                    "word_id"=> $id,
-                    "meaning"=> $x["meaning"],
-                    "explanation_of_meaning"=> $x["explanation_of_meaning"],
-                    "example"=> $x["example"],
-                    "example_meaning"=> $x["example_meaning"],
-                    "image"=> $x["image"]
-                );
-                $meaning = Meaning::create($data_meaning);
+                // Create new data related
+                foreach($arr_meanings as $x){
+                
+                    $data_meaning = array(
+                        "word_id"=> $id,
+                        "meaning"=> $x["meaning"],
+                        "explanation_of_meaning"=> $x["explanation_of_meaning"],
+                        "example"=> $x["example"],
+                        "example_meaning"=> $x["example_meaning"],
+                        "image"=> $x["image"]
+                    );
+                    $meaning = Meaning::create($data_meaning);
 
-                foreach($x["tags_id"] as $tag_id){
-                    $meaning->tags()->attach($tag_id);
+                    foreach($x["tags_id"] as $tag_id){
+                        $meaning->tags()->attach($tag_id);
+                    }
                 }
             }
+
+            if (isset($synonym)) {
+                foreach ($synonym as $value) {
+                    $word2_id = Word::where('word', $value)->first()->id;
+                    DB::table('word_relations')
+                        ->updateOrInsert([
+                            'word1_id' => $word->id,
+                            'word2_id' => $word2_id,
+                            'relation_type' => 1
+                        ]);
+                }
+            }
+            if (isset($antonym)) {
+                foreach ($antonym as $value) {
+                    $word2_id = Word::where('word', $value)->first()->id;
+                    DB::table('word_relations')
+                        ->updateOrInsert([
+                            'word1_id' => $word->id,
+                            'word2_id' => $word2_id,
+                            'relation_type' => 0
+                        ]);
+                }
+            }
+            DB::commit();
             return $this->success($word, 'Word has been updated successfully');
         } catch (QueryException $th) {
+            DB::rollBack();
             return \response()->json($th->errorInfo);
         }
     }
